Reorganise trackbar source
