+++ ui_helpers.vcxproj
@@ -171,6 +171,7 @@
     <ClInclude Include="text_drawing.h" />
     <ClInclude Include="theming.h" />
     <ClInclude Include="trackbar.h" />
+    <ClInclude Include="trackbar_base.h" />
     <ClInclude Include="win32_helpers.h" />
   </ItemGroup>
   <ItemGroup>
@@ -287,6 +288,7 @@
     </ClCompile>
     <ClCompile Include="text_drawing.cpp" />
     <ClCompile Include="trackbar.cpp" />
+    <ClCompile Include="trackbar_base.cpp" />
     <ClCompile Include="win32_helpers.cpp" />
   </ItemGroup>
   <ItemGroup>
